re-organize attributes, entities, and components
