+++ client/src/forms/FormBuilder.tsx
@@ -1,0 +1,4 @@
+import { BuilderEntities, useBuilderStore } from "@coltorapps/builder-react";
+
+import { LabelAttribute } from "./Components";
+import { textFieldEntity } from "./entities";
