+++ client/src/forms/entities.ts
@@ -2,7 +2,7 @@ import { z } from 'zod';
 
 import { createEntity } from '@coltorapps/builder';
 
-import { labelAttribute } from './label/label-attribute';
+import { labelAttribute } from './attributes';
 
 // Think of entities with attributes as components with props. 
 // For example, you can define a text field entity, and users can later add multiple instances of text fields to a form.
